# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MYG — Moroccan Youth Governance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface Brochure {
  id: number;
  title: string;
  description: string;
  /** e.g. events, services, partnerships */
  category: string;
  /** Object storage path for the PDF */
  objectPath: string;
  createdAt: string;
}

export interface BrochureInput {
  /** @minLength 1 */
  title: string;
  description: string;
  category: string;
  objectPath: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorResponse {
  error: string;
}