# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MYG — Moroccan Youth Governance API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form submissions
  - name: brochures
    description: Downloadable brochure documents
  - name: storage
    description: File storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Submits the contact form and sends an email notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Contact form submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /brochures:
    get:
      operationId: listBrochures
      tags: [brochures]
      summary: List all brochures
      description: Returns all published brochures available for download
      responses:
        "200":
          description: List of brochures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brochure"
    post:
      operationId: createBrochure
      tags: [brochures]
      summary: Create a new brochure
      description: Adds a new brochure entry with an uploaded PDF path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrochureInput"
      responses:
        "201":
          description: Brochure created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brochure"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /brochures/{id}:
    delete:
      operationId: deleteBrochure
      tags: [brochures]
      summary: Delete a brochure
      description: Removes a brochure entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brochure deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Brochure not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      description: Returns a presigned URL for direct PDF upload to object storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ContactInput:
      type: object
      required:
        - name
        - email
        - subject
        - message
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        subject:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        organization:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    ContactResult:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string

    Brochure:
      type: object
      required:
        - id
        - title
        - description
        - category
        - objectPath
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          description: "e.g. events, services, partnerships"
        objectPath:
          type: string
          description: "Object storage path for the PDF"
        createdAt:
          type: string
          format: date-time

    BrochureInput:
      type: object
      required:
        - title
        - description
        - category
        - objectPath
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        objectPath:
          type: string

    DeleteResult:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean

    UploadUrlRequest:
      type: object
      required:
        - name
        - size
        - contentType
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string

    UploadUrlResponse:
      type: object
      required:
        - uploadURL
        - objectPath
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string