# lib/api-zod/src/generated/types/brochure.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MYG — Moroccan Youth Governance API
 * OpenAPI spec version: 0.1.0
 */

export interface Brochure {
  id: number;
  title: string;
  description: string;
  /** e.g. events, services, partnerships */
  category: string;
  /** Object storage path for the PDF */
  objectPath: string;
  createdAt: Date;
}