# lib/api-zod/src/generated/types/brochureInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MYG — Moroccan Youth Governance API
 * OpenAPI spec version: 0.1.0
 */

export interface BrochureInput {
  /** @minLength 1 */
  title: string;
  description: string;
  category: string;
  objectPath: string;
}