# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MYG — Moroccan Youth Governance API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  phone?: string | null;
}